+++ lib/geordi/commands/yarn_install.rb
@@ -1,7 +1,7 @@
 desc 'yarn-install', 'Runs yarn install if required', hide: true
 
 def yarn_install
-  if File.exist?('package.json') && !system('yarn check --integrity > /dev/null 2>&1')
+  if File.exist?('yarn.lock') && !system('yarn check --integrity > /dev/null 2>&1')
     Interaction.announce 'Yarn install'
     Util.run!('yarn install')
   end